add drr_rewrite_pattern.cc
